fix (mniej więcej)
nie wiem dlaczego w removeImg() jak dam zamiast location.reload() to loadSite(site) to najpierw wykonuje loadSite(site) a potem i tak robi reload strony.

+++ admin/removeimg.php
@@ -5,8 +5,6 @@ if (!isset($_SESSION['login'])) {
 	die("Brak dostępu");
 }
 
-if ($_SERVER['REQUEST_METHOD'] === 'GET') {
-	if (isset($_GET['name'])){
-		unlink($_SERVER['DOCUMENT_ROOT'] . '/media/upload/' . $_GET['name']);
-	}
+if ($_SERVER['REQUEST_METHOD'] === 'GET' && isset($_GET['name'])) {
+	unlink($_SERVER['DOCUMENT_ROOT'] . '/media/upload/' . $_GET['name']);
 }
